Make environment detection consistent
